requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
google-generativeai==0.3.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread==6.0.0
python-dotenv==1.0.0
aiohttp==3.9.1
python-multipart==0.0.6
pyyaml==6.0.1

